+++ src/main/java/tqs/project/api/repositories/PedidoRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import tqs.project.api.models.Pedido;
 
 @Repository
-public interface PedidoRepository extends JpaRepository<Pedido, Double> {
+public interface PedidoRepository extends JpaRepository<Pedido, Long> {
     List<Pedido> findAllByStatus(int status);
 }
